Refactor: caledarHelper class로 코드 분리
calendar의 뷰를 그리는데 있어서 도움을 주는 코드를 calendarHelper class로 분리
action 함수 (Bool, Date) -> Void 로 선언 및 코드 위치 시킴
